Refactor changes into a plugin
